Check for rateErrors & update styles
