--- common/components/BalanceSidebar/Equiv.scss
@@ -1,87 +1,0 @@
-@import 'common/sass/variables';
-@import 'common/sass/mixins';
-
-.EquivalentValues {
-  &-header {
-    display: flex;
-    align-items: center;
-    margin-bottom: $space;
-
-    .Select {
-      flex-grow: 1;
-    }
-  }
-
-  &-title {
-    margin: 0;
-    margin-right: 16px;
-  }
-
-  &-values {
-    &-currency {
-      margin-bottom: $space-xs;
-      &-label {
-        white-space: pre-wrap;
-        display: inline-block;
-        min-width: 36px;
-        opacity: 0.54;
-        margin-right: 8px;
-      }
-      &-value {
-        font-weight: 600;
-        @include mono;
-      }
-    }
-  }
-
-  &-offline {
-    margin-bottom: 0;
-    text-align: center;
-  }
-}
-
-// .EquivalentValues {
-//   &-title {
-//     margin-top: 0;
-//     margin-bottom: $space;
-//   }
-
-//   &-values {
-//     list-style: none;
-//     padding: 0;
-//     @include clearfix;
-
-//     &-currency {
-//       float: left;
-//       width: 50%;
-//       margin-bottom: $space-xs;
-
-//       &:nth-child(odd) {
-//         padding-right: $space-sm;
-//       }
-//       &:nth-child(even) {
-//         padding-left: $space-sm;
-//       }
-
-//       &-label {
-//         white-space: pre-wrap;
-//         display: inline-block;
-//         min-width: 36px;
-//       }
-//       &-value {
-//         font-weight: 600;
-//         @include mono;
-//       }
-//     }
-
-//     &-loader {
-//       padding: 25px 0;
-//       text-align: center;
-//     }
-//   }
-
-//   &-offline {
-//     margin-bottom: 0;
-//     text-align: center;
-//   }
-// }
