+++ common/components/BalanceSidebar/EquivalentValues.tsx
@@ -1,177 +1,212 @@
-import * as React from 'react';
-import BN from 'bn.js';
+import React from 'react';
 import translate from 'translations';
-import { State } from 'reducers/rates';
-import { rateSymbols, TFetchCCRates } from 'actions/rates';
+import { UnitDisplay, Spinner } from 'components/ui';
+import Select from 'react-select';
+import { TFetchCCRates, rateSymbols } from 'actions/rates';
+import { chain, flatMap } from 'lodash';
+import { NetworkConfig } from 'config/data';
+import { State as RatesState } from 'reducers/rates';
 import { TokenBalance } from 'selectors/wallet';
 import { Balance } from 'libs/wallet';
-import { NetworkConfig } from 'config/data';
-import { ETH_DECIMAL, convertTokenBase } from 'libs/units';
-import Spinner from 'components/ui/Spinner';
-import UnitDisplay from 'components/ui/UnitDisplay';
 import './EquivalentValues.scss';
 
-const ALL_OPTION = 'All';
+const isFiat = (rate: string): boolean => {
+  return rateSymbols.slice(0, 4).includes(rate as any);
+};
+
+interface DefaultOption {
+  label: string;
+  value: any;
+}
+
+interface Option {
+  label: string;
+  value: string;
+}
+
+interface State {
+  equivalentValues: Option | DefaultOption;
+  options: Option[];
+}
 
 interface Props {
-  balance?: Balance;
-  tokenBalances?: TokenBalance[];
-  rates: State['rates'];
-  ratesError?: State['ratesError'];
+  balance: Balance;
+  tokenBalances: TokenBalance[];
+  rates: RatesState['rates'];
   fetchCCRates: TFetchCCRates;
+  ratesError: RatesState['ratesError'];
   network: NetworkConfig;
-  isOffline: boolean;
+  offline: boolean;
 }
 
-interface CmpState {
-  currency: string;
-}
-
-export default class EquivalentValues extends React.Component<Props, CmpState> {
-  public state = {
-    currency: ALL_OPTION
-  };
-  private balanceLookup: { [key: string]: Balance['wei'] | undefined } = {};
-  private decimalLookup: { [key: string]: number } = {};
+class Equiv extends React.Component<Props, State> {
   private requestedCurrencies: string[] | null = null;
-
   public constructor(props: Props) {
     super(props);
-    this.makeBalanceLookup(props);
+    const { balance, tokenBalances, network } = this.props;
+    this.state = {
+      equivalentValues: this.defaultOption(balance, tokenBalances, network),
+      options: []
+    };
 
     if (props.balance && props.tokenBalances) {
       this.fetchRates(props);
     }
   }
 
+  public defaultOption(balance, tokenBalances, network): DefaultOption {
+    return {
+      label: 'All',
+      value: [{ symbol: network.unit, balance: balance.wei }, ...tokenBalances]
+    };
+  }
+
   public componentWillReceiveProps(nextProps: Props) {
-    const { balance, tokenBalances, isOffline } = this.props;
+    const { balance, tokenBalances, offline } = this.props;
     if (
       nextProps.balance !== balance ||
       nextProps.tokenBalances !== tokenBalances ||
-      nextProps.isOffline !== isOffline
+      nextProps.offline !== offline
     ) {
-      this.makeBalanceLookup(nextProps);
+      const defaultOption = this.defaultOption(
+        nextProps.balance,
+        nextProps.tokenBalances,
+        nextProps.network
+      );
+      const options: Option[] = [
+        defaultOption,
+        { label: nextProps.network.unit, value: nextProps.balance.wei },
+        ...Object.values(nextProps.tokenBalances).map(token => {
+          return { label: token.symbol, value: token.balance };
+        })
+      ];
+      const equivalentValues =
+        options.find(opt => opt.label === this.state.equivalentValues.label) || defaultOption;
+      this.setState({
+        equivalentValues,
+        options
+      });
       this.fetchRates(nextProps);
     }
   }
 
-  public render() {
-    const { balance, tokenBalances, rates, ratesError, isOffline, network } = this.props;
-    const { currency } = this.state;
+  public selectOption = equivalentValues => {
+    this.setState({ equivalentValues });
+  };
 
-    // There are a bunch of reasons why the incorrect balances might be rendered
-    // while we have incomplete data that's being fetched.
+  public render(): JSX.Element {
+    const { balance, offline, tokenBalances, rates, network, ratesError } = this.props;
+    const { equivalentValues, options } = this.state;
     const isFetching =
       !balance || balance.isPending || !tokenBalances || Object.keys(rates).length === 0;
-    // Currency exists in rates or the all option is selected
-    const rateExistsOrAll = rates[currency] || currency === ALL_OPTION;
 
-    let valuesEl;
-    if (!isFetching && rateExistsOrAll && !network.isTestnet) {
-      const values = this.getEquivalentValues(currency);
-      valuesEl = rateSymbols.map(key => {
-        if (!values[key] || key === currency) {
-          return null;
-        }
-
-        return (
-          <li className="EquivalentValues-values-currency" key={key}>
-            <span className="EquivalentValues-values-currency-label">{key}:</span>{' '}
-            <span className="EquivalentValues-values-currency-value">
-              <UnitDisplay
-                unit={'ether'}
-                value={values[key]}
-                displayShortBalance={3}
-                checkOffline={true}
-              />
-            </span>
-          </li>
-        );
-      });
-    } else if (network.isTestnet) {
-      valuesEl = (
-        <div className="text-center">
-          <h5 style={{ color: 'red' }}>
-            On test network, equivalent values will not be displayed.
-          </h5>
-        </div>
-      );
-    } else if (ratesError) {
-      valuesEl = <h5>{ratesError}</h5>;
-    } else if (tokenBalances && tokenBalances.length === 0) {
-      valuesEl = <h5>No tokens found!</h5>;
-    } else {
-      valuesEl = (
-        <div className="EquivalentValues-values-loader">
-          <Spinner size="x3" />
-        </div>
-      );
-    }
+    const Value = ({ rate, value }) => (
+      <div className="EquivalentValues-values-currency">
+        <span className="EquivalentValues-values-currency-label">{rate}</span>{' '}
+        <span className="EquivalentValues-values-currency-value">
+          <UnitDisplay
+            unit={'ether'}
+            value={value}
+            displayShortBalance={isFiat(rate) ? 2 : 3}
+            checkOffline={true}
+          />
+        </span>
+      </div>
+    );
 
     return (
       <div className="EquivalentValues">
-        <h5 className="EquivalentValues-title">
-          {translate('sidebar_Equiv')} for{' '}
-          <select
-            className="EquivalentValues-title-symbol"
-            onChange={this.changeCurrency}
-            value={currency}
-          >
-            <option value={ALL_OPTION}>All Tokens</option>
-            <option value="ETH">ETH</option>
-            {tokenBalances &&
-              tokenBalances.map(tk => {
-                if (!tk.balance || tk.balance.isZero() || !rates[tk.symbol]) {
-                  return;
-                }
-                const sym = tk.symbol;
-                return (
-                  <option key={sym} value={sym}>
-                    {sym}
-                  </option>
-                );
-              })}
-          </select>
-        </h5>
+        <div className="EquivalentValues-header">
+          <h5 className="EquivalentValues-title">{translate('sidebar_Equiv')}</h5>
+          <Select
+            name="equivalentValues"
+            value={equivalentValues}
+            options={options}
+            onChange={this.selectOption}
+            clearable={false}
+            searchable={false}
+          />
+        </div>
 
-        {isOffline ? (
+        {offline ? (
           <div className="EquivalentValues-offline well well-sm">
             Equivalent values are unavailable offline
           </div>
+        ) : network.isTestnet ? (
+          <div className="text-center">
+            <h5 style={{ color: 'red' }}>
+              On test network, equivalent values will not be displayed.
+            </h5>
+          </div>
+        ) : ratesError ? (
+          <h5>{ratesError}</h5>
+        ) : isFetching ? (
+          <Spinner size="x2" />
+        ) : tokenBalances && tokenBalances.length === 0 ? (
+          <h5>No tokens found!</h5>
         ) : (
-          <ul className="EquivalentValues-values">{valuesEl}</ul>
+          <div className="EquivalentValues-values">
+            {this.generateValues(equivalentValues.label, equivalentValues.value).map((equiv, i) => (
+              <Value rate={equiv.rate} value={equiv.value} key={i} />
+            ))}
+          </div>
         )}
       </div>
     );
   }
 
-  private changeCurrency = (ev: React.FormEvent<HTMLSelectElement>) => {
-    const currency = ev.currentTarget.value;
-    this.setState({ currency });
+  private generateValues = (unit: string, balance) => {
+    const { rates } = this.props;
+    if (unit === 'All') {
+      const allRates = Object.values(balance).map(
+        value => !!rates[value.symbol] && rates[value.symbol]
+      );
+      const allEquivalentValues = allRates.map((rateType, i) => {
+        return {
+          symbol: Object.keys(rates)[i],
+          equivalentValues: [
+            ...Object.keys(rateType).map(rate => {
+              const value =
+                balance[i] && !!balance[i].balance ? balance[i].balance.muln(rateType[rate]) : null;
+              return { rate, value };
+            })
+          ]
+        };
+      });
+      // flatten all equivalent values for each unit (ETH, ETC, OMG, etc...) into an array
+      const collection = flatMap([
+        ...Object.values(allEquivalentValues).map(v => v.equivalentValues)
+      ]);
+      // group equivalent values by rate (USD, EUR, etc...)
+      const groupedCollection = chain(collection)
+        .groupBy('rate')
+        .mapValues(v => Object.values(v).map(s => s.value))
+        .value();
+      // finally, add all the equivalent values together and return an array of objects with the sum of equivalent values for each rate
+      return Object.values(groupedCollection).map((v, i) => {
+        return {
+          rate: Object.keys(groupedCollection)[i],
+          value: v.reduce((acc, curr) => acc && curr && acc.add(curr))
+        };
+      });
+    }
+    const ratesObj = { ...rates[unit] };
+    const equivValues = Object.keys(ratesObj).map(key => {
+      const value = balance !== null ? balance.muln(ratesObj[key]) : null;
+      return { rate: key, value };
+    });
+    return equivValues;
   };
 
-  private makeBalanceLookup(props: Props) {
-    const tokenBalances = props.tokenBalances || [];
-    this.balanceLookup = tokenBalances.reduce(
-      (prev, tk) => {
-        // Piggy-back off of this reduce to add to decimal lookup
-        this.decimalLookup[tk.symbol] = tk.decimal;
-        prev[tk.symbol] = tk.balance;
-        return prev;
-      },
-      { ETH: props.balance && props.balance.wei }
-    );
-  }
-
   private fetchRates(props: Props) {
+    const { balance, tokenBalances, offline, fetchCCRates } = props;
     // Duck out if we haven't gotten balances yet, or we're not going to
-    if (!props.balance || !props.tokenBalances || props.isOffline) {
+    if (!balance || !tokenBalances || offline) {
       return;
     }
 
     // First determine which currencies we're asking for
-    const currencies = props.tokenBalances
+    const currencies = tokenBalances
       .filter(tk => !tk.balance.isZero())
       .map(tk => tk.symbol)
       .sort();
@@ -182,45 +217,9 @@ export default class EquivalentValues extends React.Component<Props, CmpState> {
     }
 
     // Fire off the request and save the currencies requested
-    this.props.fetchCCRates(currencies);
+    fetchCCRates(currencies);
     this.requestedCurrencies = currencies;
   }
-
-  private getEquivalentValues(
-    currency: string
-  ): {
-    [key: string]: BN | undefined;
-  } {
-    // Recursively call on all currencies
-    if (currency === ALL_OPTION) {
-      return [this.props.network.unit].concat(this.requestedCurrencies || []).reduce(
-        (prev, curr) => {
-          const currValues = this.getEquivalentValues(curr);
-          rateSymbols.forEach(sym => (prev[sym] = prev[sym].add(currValues[sym] || new BN(0))));
-          return prev;
-        },
-        rateSymbols.reduce((prev, sym) => {
-          prev[sym] = new BN(0);
-          return prev;
-        }, {})
-      );
-    }
-
-    // Calculate rates for a single currency
-    const { rates } = this.props;
-    const balance = this.balanceLookup[currency];
-    if (!balance || !rates[currency]) {
-      return {};
-    }
-
-    // Tokens with non-ether like decimals need to be adjusted to match
-    const decimal =
-      this.decimalLookup[currency] === undefined ? ETH_DECIMAL : this.decimalLookup[currency];
-    const adjustedBalance = convertTokenBase(balance, decimal, ETH_DECIMAL);
-
-    return rateSymbols.reduce((prev, sym) => {
-      prev[sym] = adjustedBalance.muln(rates[currency][sym]);
-      return prev;
-    }, {});
-  }
 }
+
+export default Equiv;
