+++ common/components/BalanceSidebar/index.tsx
@@ -7,10 +7,9 @@ import { AppState } from 'reducers';
 import { getNetworkConfig } from 'selectors/config';
 import { getShownTokenBalances, getWalletInst, TokenBalance } from 'selectors/wallet';
 import AccountInfo from './AccountInfo';
-import EquivalentValues from './EquivalentValues';
 import Promos from './Promos';
 import TokenBalances from './TokenBalances';
-import Equiv from 'components/BalanceSidebar/Equiv';
+import EquivalentValues from 'components/BalanceSidebar/EquivalentValues';
 
 interface Props {
   wallet: IWallet;
@@ -54,24 +53,15 @@ export class BalanceSidebar extends React.Component<Props, {}> {
       {
         name: 'Equivalent Values',
         content: (
-          <Equiv
+          <EquivalentValues
             offline={isOffline}
             rates={rates}
             balance={balance}
             network={network}
             tokenBalances={tokenBalances}
-            // ratesError={ratesError}
+            ratesError={ratesError}
             fetchCCRates={this.props.fetchCCRates}
           />
-          // <EquivalentValues
-          //   network={network}
-          //   balance={balance}
-          //   tokenBalances={tokenBalances}
-          //   rates={rates}
-          //   ratesError={ratesError}
-          //   fetchCCRates={this.props.fetchCCRates}
-          //   isOffline={isOffline}
-          // />
         )
       }
     ];
